Moved classes Segment and Segments into separate files
