+++ main/core/src/mill/define/Segment.scala
@@ -1,0 +1,13 @@
+package mill.define
+
+sealed trait Segment {
+  def pathSegments: Seq[String] = this match {
+    case Segment.Label(s) => Seq(s)
+    case Segment.Cross(vs) => vs.map(_.toString)
+  }
+}
+
+object Segment {
+  final case class Label(value: String) extends Segment
+  final case class Cross(value: Seq[Any]) extends Segment
+}
